backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-green20220302==2.20.0
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_util==0.3.13
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==44.0.3
darabonba-core==1.0.3
Django==5.2.1
django-cors-headers==4.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
